Add top-level subpackage imports in indexed and interventional modules (#169)
* move cf handlers out of init

* add imports
