+++ causal_pyro/indexed/__init__.py
@@ -1,0 +1,1 @@
+from . import internals  # noqa: F401
