+++ causal_pyro/interventional/__init__.py
@@ -1,0 +1,1 @@
+from . import handlers  # noqa: F401
